builder_api: move GetExecutionPayloadBidResponse to combined
