+++ builder_api/src/combined.rs
@@ -8,6 +8,7 @@ use types::{
     combined::{ExecutionPayload, ExecutionPayloadHeader},
     deneb::primitives::KzgCommitment,
     electra::containers::ExecutionRequests,
+    gloas::containers::SignedExecutionPayloadBid,
     nonstandard::{KzgProofs, Phase, WithBlobsAndMev},
     phase0::primitives::Uint256,
     preset::Preset,
@@ -288,6 +289,44 @@ impl<P: Preset> ExecutionPayloadAndBlobsBundle<P> {
             Self::Deneb(_) => Phase::Deneb,
             Self::Electra(_) => Phase::Electra,
             Self::Fulu(_) => Phase::Fulu,
+        }
+    }
+}
+
+#[derive(Debug, Deserialize)]
+#[serde(
+    bound = "",
+    deny_unknown_fields,
+    rename_all = "lowercase",
+    tag = "version",
+    content = "data"
+)]
+pub enum GetExecutionPayloadBidResponse<P: Preset> {
+    Gloas(SignedExecutionPayloadBid<P>),
+}
+
+impl<P: Preset> SszSize for GetExecutionPayloadBidResponse<P> {
+    const SIZE: Size = SignedExecutionPayloadBid::<P>::SIZE;
+}
+
+impl<P: Preset> SszRead<Phase> for GetExecutionPayloadBidResponse<P> {
+    fn from_ssz_unchecked(phase: &Phase, bytes: &[u8]) -> Result<Self, ReadError> {
+        // Container is Gloas-shaped; header may report any post-Gloas phase.
+        if *phase >= Phase::Gloas {
+            return Ok(Self::Gloas(SszReadDefault::from_ssz_default(bytes)?));
+        }
+
+        Err(ReadError::Custom {
+            message: "execution payload bid response is only available from Gloas onwards",
+        })
+    }
+}
+
+impl<P: Preset> GetExecutionPayloadBidResponse<P> {
+    #[must_use]
+    pub fn into_bid(self) -> SignedExecutionPayloadBid<P> {
+        match self {
+            Self::Gloas(bid) => bid,
         }
     }
 }
